Update on "Adding a version serialization type to ConvPackedParam"
This PR changes the format of `ConvPackedParam` in a nearly backwards-compatible way:
* a new format is introduced which has more flexibility and a lower on-disk size
* custom pickle functions are added to `ConvPackedParams` which know how to load the old format
* the custom pickle functions are **not** BC because the output type of `__getstate__` has changed.  We expect this to be acceptable as no user flows are actually broken (loading a v1 model with v2 code works), which is why we whitelist the failure.

Test plan (TODO finalize):
* verify that saving and loading models with https://gist.github.com/vkuzo/f3616c5de1b3109cb2a1f504feed69be works
* TODO before land: add test coverage for loading v1 models
* TODO before land: verify numerics do not change
* TODO before land: verify all combinations of (trace, script) x (bias, no_bias) work


Note: this is a newer copy of https://github.com/pytorch/pytorch/pull/40003

[ghstack-poisoned]

+++ torch/csrc/jit/passes/onnx/unpack_quantized_weights.cpp
@@ -190,7 +190,7 @@ void unpackQuantizedWeightsHelper(
 
         const int64_t kSpatialDim = conv_params_packed[0].item<int64_t>();
         // skip kSpatialDim
-        int idx = 1;
+        int64_t idx = 1;
         for (; idx < kSpatialDim + 1; ++idx) {
           stride_int.emplace_back(conv_params_packed[idx].item<int64_t>());
         }
